Removed empty Model classes and renamed existing Model classes.
